[OPIK-6739] [FE] refactor: share trace filter-chip definitions
Move the trace chip definitions, their display order and their default
pinned set out of the Logs page so any surface showing traces can consume
them. The entity-scoped trace logs view adopts them next.

They land under pages-shared/traces rather than shared/filter-chips because
the option sources (tags, metadata paths, error types) are pages-shared
hooks, and the no-shared-importing-pages rule forbids shared/ from reaching
them.

The dynamic-chip builder gains a logsSource parameter — it hard-coded the
SDK source, which is wrong for evaluator traces — and the Logs page imports
it back for its span chips. Logs page behaviour is unchanged.
